requirements.txt
slackclient
python-dotenv
slackeventsapi
flask
groq